Implemented Java test client using JacORB
The CORBA cpp server can now be connected to from a Java client which
allows some interactive testing of the toaster model. The functionality
for toasting bread has also been updated to be asynchronous. The main
functionality now executes in a separate thread so other operations may
be carried out concurrently.